Refactor: 파일 기능별로 분리 #SCRUM-65
